StandaloneTestStrategy: add lcov_merger to tools of TestRunner spawns
The default lcov_merger (@bazel_tools//tools/test:lcov_merger) is a
java_binary that requires runfiles to work.
In StandaloneTestStrategy, it was added to the inputs of TestRunner
spawns, but not to tools.
This prevents the runfiles from being available to the action,
preventing coverage to be collected during remote execution.

Fixes #4033